Use score similar to problem definition (with assumption that best solution has 0 energy).

+++ src/cpp/solver.h
@@ -7,6 +7,6 @@ class Solver
 {
 public:
     static uint64_t Solve(const Matrix& m, Trace& output);
-    static double Solve(unsigned model_index);
+    static unsigned Solve(unsigned model_index);
     static void SolveAll();
 };
